refactor(cost): close round-6 review polish
Six 5-minute fixes from the deep review:

- BookedCost::new test: add `booked_cost_new_panics_in_debug_on_overflow`
  pinning the debug-assert message names overflow. Without it, a future
  refactor of check_invariant's Err path could swallow the overflow case
  at the `new` call site (try_new is covered, new wasn't).
- Validator destructure: one-line comment noting the
  `BookedCost { total: value, .. }` pattern is read-only; future
  contributors must not copy it to construct (would bypass the
  invariant enforced by new/try_new).
- .cargo/config.toml: warn about the RUSTFLAGS env-var-vs-config
  precedence trap. Env wins outright, no merge — any future CI step
  exporting RUSTFLAGS during a wasm build will silently drop the
  getrandom_backend cfg and reproduce the original failure.
- Cache byte-fixture test: moved from rustledger-core::cost to
  rustledger-loader::cache where CACHE_VERSION is in scope. Added
  FIXTURE_VERSION tripwire that fires if CACHE_VERSION advances past
  8 without the fixtures being regenerated. Gated to
  `cfg(target_endian = "little")` since rkyv to_bytes uses native
  endianness — non-portable bytes are fine for same-machine cache
  reads but would spuriously fail tests on big-endian platforms.
